requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.7.0
django-braces==1.17.0
django-ckeditor-5==0.2.17
django-etc==1.4.0
django-hitcount==1.3.5
django-model-utils==5.0.0
django-star-ratings==0.9.2
django-taggit==6.1.0
dotenv==0.9.9
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
swapper==1.4.0
typing_extensions==4.13.0
upgrade-requirements==1.7.0
urllib3==2.4.0
whitenoise==6.9.0
